// src/app/_services/model/ppob.ts
export class Ppob {

    PPOBID: number;
    GSTNumber: string;
    CompanyID: number;
    CountryID: number;
    StateID: number;
    StateName:string;

    City: string;
    Address1: string;
    Address2: string;
    Pincode: string;
    ContactName: string;
    ContactNumber: string;

    Email: string; 
    IsActive: boolean;
    GSTFilePath: string;
    GSTPortalUserName: string;

    GSTPortalPassword: string;
    Action: string;
    LoginId: number;
}